refactor: move cache handling to common superclass
