+++ src/language/builtins/safe-ds-annotations.ts
@@ -1,20 +1,11 @@
-import { SafeDsServices } from '../safe-ds-module.js';
 import { resolveRelativePathToBuiltinFile } from './fileFinder.js';
-import { isSdsAnnotation, isSdsModule, SdsAnnotatedObject, SdsAnnotation } from '../generated/ast.js';
-import { LangiumDocuments, WorkspaceCache } from 'langium';
-import { annotationCallsOrEmpty, moduleMembersOrEmpty } from '../helpers/nodeProperties.js';
+import { isSdsAnnotation, SdsAnnotatedObject, SdsAnnotation } from '../generated/ast.js';
+import { annotationCallsOrEmpty } from '../helpers/nodeProperties.js';
+import { SafeDsModuleMembers } from './safe-ds-module-members.js';
 
 const CORE_ANNOTATIONS_URI = resolveRelativePathToBuiltinFile('safeds/lang/coreAnnotations.sdsstub');
 
-export class SafeDsAnnotations {
-    private readonly langiumDocuments: LangiumDocuments;
-    private readonly cache: WorkspaceCache<string, SdsAnnotation>;
-
-    constructor(services: SafeDsServices) {
-        this.langiumDocuments = services.shared.workspace.LangiumDocuments;
-        this.cache = new WorkspaceCache(services.shared);
-    }
-
+export class SafeDsAnnotations extends SafeDsModuleMembers<SdsAnnotation> {
     isDeprecated(node: SdsAnnotatedObject | undefined): boolean {
         return annotationCallsOrEmpty(node).some((it) => {
             const annotation = it.annotation?.ref;
@@ -38,29 +29,6 @@ export class SafeDsAnnotations {
     }
 
     private getAnnotation(name: string): SdsAnnotation | undefined {
-        if (this.cache.has(name)) {
-            return this.cache.get(name);
-        }
-
-        if (!this.langiumDocuments.hasDocument(CORE_ANNOTATIONS_URI)) {
-            /* c8 ignore next 2 */
-            return undefined;
-        }
-
-        const document = this.langiumDocuments.getOrCreateDocument(CORE_ANNOTATIONS_URI);
-        const root = document.parseResult.value;
-        if (!isSdsModule(root)) {
-            /* c8 ignore next 2 */
-            return undefined;
-        }
-
-        const firstMatchingModuleMember = moduleMembersOrEmpty(root).find((m) => m.name === name);
-        if (!isSdsAnnotation(firstMatchingModuleMember)) {
-            /* c8 ignore next 2 */
-            return undefined;
-        }
-
-        this.cache.set(name, firstMatchingModuleMember);
-        return firstMatchingModuleMember;
+        return this.getModuleMember(CORE_ANNOTATIONS_URI, name, isSdsAnnotation);
     }
 }
