+++ src/language/builtins/safe-ds-module-members.ts
@@ -1,0 +1,43 @@
+import { SafeDsServices } from '../safe-ds-module.js';
+import { isSdsModule, SdsModuleMember } from '../generated/ast.js';
+import { LangiumDocuments, URI, WorkspaceCache } from 'langium';
+import { moduleMembersOrEmpty } from '../helpers/nodeProperties.js';
+
+export abstract class SafeDsModuleMembers<T extends SdsModuleMember> {
+    private readonly langiumDocuments: LangiumDocuments;
+    private readonly cache: WorkspaceCache<string, T>;
+
+    constructor(services: SafeDsServices) {
+        this.langiumDocuments = services.shared.workspace.LangiumDocuments;
+        this.cache = new WorkspaceCache(services.shared);
+    }
+
+    protected getModuleMember(uri: URI, name: string, predicate: (node: unknown) => node is T): T | undefined {
+        const key = `${uri.toString()}#${name}`;
+
+        if (this.cache.has(key)) {
+            return this.cache.get(name);
+        }
+
+        if (!this.langiumDocuments.hasDocument(uri)) {
+            /* c8 ignore next 2 */
+            return undefined;
+        }
+
+        const document = this.langiumDocuments.getOrCreateDocument(uri);
+        const root = document.parseResult.value;
+        if (!isSdsModule(root)) {
+            /* c8 ignore next 2 */
+            return undefined;
+        }
+
+        const firstMatchingModuleMember = moduleMembersOrEmpty(root).find((m) => m.name === name);
+        if (!predicate(firstMatchingModuleMember)) {
+            /* c8 ignore next 2 */
+            return undefined;
+        }
+
+        this.cache.set(name, firstMatchingModuleMember);
+        return firstMatchingModuleMember;
+    }
+}
